#15 : /api/odbcinfo has been granted query specification.

+++ src/restapi/ApiOdbcInfos.cpp
@@ -1,3 +1,5 @@
+#include <windows.h>
+#include <shlwapi.h>
 #include <tchar.h>
 #include "ApiOdbcInfos.h"
 #include "dataaccess.h"
@@ -7,44 +9,48 @@
 
 StkObject* ApiOdbcInfos::Execute(StkObject* ReqObj, int Method, TCHAR UrlPath[128], int* ResultCode, TCHAR Locale[3])
 {
-	SQLTCHAR ConnStrPostgreSql[Global::MAX_PARAM_LENGTH];
-	SQLTCHAR ConnStrMariaDb[Global::MAX_PARAM_LENGTH];
-	SQLTCHAR ConnStrMySql[Global::MAX_PARAM_LENGTH];
-
-	DbAccessor* DaPostgreSql = NULL;
-	DbAccessor* DaMariaDb    = NULL;
-	DbAccessor* DaMySql      = NULL;
-
-	DaPostgreSql = OdbcManager::GetInstance()->CreateAccessorObject(OdbcManager::POSTGRESQL_ACCESSOR);
-	DaMariaDb    = OdbcManager::GetInstance()->CreateAccessorObject(OdbcManager::MARIADB_ACCESSOR);
-	DaMySql      = OdbcManager::GetInstance()->CreateAccessorObject(OdbcManager::MYSQL_ACCESSOR);
-
-	DaPostgreSql->GetDefaultConnStr(ConnStrPostgreSql);
-	DaMariaDb->GetDefaultConnStr(ConnStrMariaDb);
-	DaMySql->GetDefaultConnStr(ConnStrMySql);
-
-	OdbcManager::GetInstance()->DeleteAccessorObject(DaPostgreSql);
-	OdbcManager::GetInstance()->DeleteAccessorObject(DaMariaDb);
-	OdbcManager::GetInstance()->DeleteAccessorObject(DaMySql);
-
 	StkObject* ResObj = new StkObject(_T(""));
 	AddCodeAndMsg(ResObj, 0, _T(""), _T(""));
 	StkObject* DatObj = new StkObject(_T("Data"));
 
-	StkObject* DatObjPostgreSql = new StkObject(_T("OdbcInfo"));
-	DatObjPostgreSql->AppendChildElement(new StkObject(_T("DbType"), _T("PostgreSQL")));
-	DatObjPostgreSql->AppendChildElement(new StkObject(_T("ConnStr"), ConnStrPostgreSql));
-	DatObj->AppendChildElement(DatObjPostgreSql);
+	if (StrStr(UrlPath, _T("?query=default"))) {
+		SQLTCHAR ConnStrPostgreSql[Global::MAX_PARAM_LENGTH];
+		SQLTCHAR ConnStrMariaDb[Global::MAX_PARAM_LENGTH];
+		SQLTCHAR ConnStrMySql[Global::MAX_PARAM_LENGTH];
 
-	StkObject* DatObjMariaDbSql = new StkObject(_T("OdbcInfo"));
-	DatObjMariaDbSql->AppendChildElement(new StkObject(_T("DbType"), _T("MariaDB")));
-	DatObjMariaDbSql->AppendChildElement(new StkObject(_T("ConnStr"), ConnStrMariaDb));
-	DatObj->AppendChildElement(DatObjMariaDbSql);
+		DbAccessor* DaPostgreSql = NULL;
+		DbAccessor* DaMariaDb    = NULL;
+		DbAccessor* DaMySql      = NULL;
 
-	StkObject* DatObjMySql = new StkObject(_T("OdbcInfo"));
-	DatObjMySql->AppendChildElement(new StkObject(_T("DbType"), _T("MySQL")));
-	DatObjMySql->AppendChildElement(new StkObject(_T("ConnStr"), ConnStrMySql));
-	DatObj->AppendChildElement(DatObjMySql);
+		DaPostgreSql = OdbcManager::GetInstance()->CreateAccessorObject(OdbcManager::POSTGRESQL_ACCESSOR);
+		DaMariaDb    = OdbcManager::GetInstance()->CreateAccessorObject(OdbcManager::MARIADB_ACCESSOR);
+		DaMySql      = OdbcManager::GetInstance()->CreateAccessorObject(OdbcManager::MYSQL_ACCESSOR);
+
+		DaPostgreSql->GetDefaultConnStr(ConnStrPostgreSql);
+		DaMariaDb->GetDefaultConnStr(ConnStrMariaDb);
+		DaMySql->GetDefaultConnStr(ConnStrMySql);
+
+		OdbcManager::GetInstance()->DeleteAccessorObject(DaPostgreSql);
+		OdbcManager::GetInstance()->DeleteAccessorObject(DaMariaDb);
+		OdbcManager::GetInstance()->DeleteAccessorObject(DaMySql);
+
+		StkObject* DatObjPostgreSql = new StkObject(_T("OdbcInfo"));
+		DatObjPostgreSql->AppendChildElement(new StkObject(_T("DbType"), _T("PostgreSQL")));
+		DatObjPostgreSql->AppendChildElement(new StkObject(_T("ConnStr"), ConnStrPostgreSql));
+		DatObj->AppendChildElement(DatObjPostgreSql);
+
+		StkObject* DatObjMariaDbSql = new StkObject(_T("OdbcInfo"));
+		DatObjMariaDbSql->AppendChildElement(new StkObject(_T("DbType"), _T("MariaDB")));
+		DatObjMariaDbSql->AppendChildElement(new StkObject(_T("ConnStr"), ConnStrMariaDb));
+		DatObj->AppendChildElement(DatObjMariaDbSql);
+
+		StkObject* DatObjMySql = new StkObject(_T("OdbcInfo"));
+		DatObjMySql->AppendChildElement(new StkObject(_T("DbType"), _T("MySQL")));
+		DatObjMySql->AppendChildElement(new StkObject(_T("ConnStr"), ConnStrMySql));
+		DatObj->AppendChildElement(DatObjMySql);
+	}
+	if (StrStr(UrlPath, _T("?query=configured"))) {
+	}
 
 	ResObj->AppendChildElement(DatObj);
